modified:   app/Http/Controllers/AssessmentController.php 	new file:   app/Http/Controllers/SawController.php 	modified:   app/Http/Controllers/StudentController.php 	modified:   app/Http/Controllers/SubCriteriaController.php 	new file:   app/Http/Controllers/TopsisController.php 	modified:   app/Models/Assessment.php 	modified:   app/Models/Student.php 	modified:   database/migrations/2023_12_03_180323_create_students_table.php 	modified:   database/migrations/2023_12_03_180403_create_criterias_table.php 	modified:   database/migrations/2023_12_03_180439_create_sub_criterias_table.php 	modified:   database/migrations/2023_12_03_181656_create_assessments_table.php 	modified:   resources/views/criterias/create.blade.php 	modified:   resources/views/criterias/edit.blade.php 	modified:   resources/views/criterias/index.blade.php 	modified:   resources/views/groups/index.blade.php 	new file:   resources/views/home.blade.php 	modified:   resources/views/layouts/admin.blade.php 	new file:   resources/views/matrixes/saw/decision.blade.php 	new file:   resources/views/matrixes/saw/normalization.blade.php 	new file:   resources/views/matrixes/saw/optimization.blade.php 	new file:   resources/views/matrixes/topsis/normalization.blade.php 	new file:   resources/views/matrixes/topsis/weightrating.blade.php 	new file:   resources/views/saw.blade.php 	new file:   resources/views/students/assessment.blade old.php 	modified:   resources/views/students/assessment.blade.php 	modified:   resources/views/students/index.blade.php 	modified:   resources/views/subcriterias/create.blade.php 	modified:   resources/views/subcriterias/edit.blade.php 	modified:   resources/views/subcriterias/index.blade.php 	new file:   resources/views/topsis.blade.php 	new file:   resources/views/values.blade.php 	deleted:    resources/views/welcome.blade.php 	modified:   routes/web.php

+++ database/migrations/2023_12_03_180403_create_criterias_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('criterias', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('weight');
+            $table->float('weight');
             $table->timestamps();
         });
     }
